block: move blockinfo out as separate module
