+++ tests/test_blockinfo.py
@@ -1,7 +1,7 @@
 # Copyright (c) 2023, Oracle and/or its affiliates.
 # Licensed under the Universal Permissive License v 1.0 as shown at https://oss.oracle.com/licenses/upl/
-from drgn_tools import block
+from drgn_tools import blockinfo
 
 
 def test_blockinfo(prog):
-    block.print_block_devs_info(prog)
+    blockinfo.print_block_devs_info(prog)
